pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "eu-multi-tenant-platform"
version = "1.0.0"
description = "EU-Grade Multi-Tenant Cloud Platform — Cost-Aware, GDPR-Native, Production-Grade SaaS"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [{name = "Ali"}]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "testcontainers[postgres,redis]>=3.7.0",
    "schemathesis>=3.25.0",
    "locust>=2.23.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
    "bandit>=1.7.7",
    "safety>=3.0.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["E501", "B008", "RUF001", "RUF003", "RUF012", "TCH003"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require database)",
    "contract: API contract tests",
    "load: Load/performance tests",
    "slow: Tests that take > 5 seconds",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
